Converge Chennai onto shared [cityId] components via capability flags
Route Chennai through the shared dynamic route instead of its flat forks,
and express every city difference as a declarative capability flag on
PlaceConfig rather than a forked component or a cityId conditional. This
is the discipline for scaling pan-India: a feature is one shared component,
a city is config + data + flags.

- Remove the /chennai -> / redirect so /chennai/* resolves.
- Add capability-flag config (dashboard, facts, flood, waterBodies,
  reservoirDataSource, historyUnit) following the existing groundwaterViews
  / heroMode pattern; centralize legacy Chennai data-file names in
  src/lib/cities/data-paths.ts.
- Facts: shared page runs the live+derived+static pipeline when
  config.facts.dynamicPipeline is set, else the static JSON snapshot.
- Dashboard: shared loaders gain a legacy-v1 source (Chennai's
  reservoir_daily / Mcft); the AI briefing, catchment context, and
  groundwater snapshot render as sections gated by config.dashboard flags.
- About: Chennai flows through the shared CityAboutContent shell via a new
  chennai-page-descriptions module (matching madurai/bangalore).
- Flood: renderer chosen by config.flood.variant; the interactive map is
  parametrized by cityId (no chennai-* literals).
- Water bodies: rich renderer selected by config.waterBodies flags and
  parametrized by cityId.

Flat Chennai routes still render (single source of truth) during the
namespace migration; they are removed + redirected in a later step.

Spec: docs/specs/multi-city-component-discipline.md
